+++ src/io/RamFileManager.hpp
@@ -14,7 +14,7 @@ class RamFileManager : public FileManager {
     };
 
   public:
-    std::unique_ptr<InputStream> getInputStream(const std::filesystem::path &filename, std::size_t offset) override;
+    std::unique_ptr<InputStream> getInputStream(const std::filesystem::path &filename, std::uint64_t offset) override;
 
     std::unique_ptr<OutputStream> getOutputStream(const std::filesystem::path &filename, bool append) override;
